+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Changed Search & Clear Icons
+
 ## [0.0.2] - 2021-09-28 - kfcromania
 
 ### Added